create client context
